Refactor fermionic mapper imports
