+++ qiskit_nature/second_quantization/operators/fermionic/mappers/jordan_wigner_mapper.py
@@ -17,13 +17,12 @@ import numpy as np
 from qiskit.opflow import PauliSumOp
 from qiskit.quantum_info.operators import Pauli
 
-from qiskit_nature.operators.second_quantization import FermionicOp
+from ..fermionic_op import FermionicOp
 
-from .fermionic_mapper import FermionicMapper
-from .qubit_mapper import QubitMapper
+from ...qubit_mapper import QubitMapper
 
 
-class JordanWignerMapper(FermionicMapper):  # pylint: disable=missing-class-docstring
+class JordanWignerMapper(QubitMapper):  # pylint: disable=missing-class-docstring
     def __init__(self):
         """The Jordan-Wigner fermion-to-qubit mapping."""
         super().__init__(allows_two_qubit_reduction=False)
