+++ qiskit_nature/second_quantization/operators/fermionic/mappers/parity_mapper.py
@@ -18,13 +18,12 @@ import numpy as np
 from qiskit.opflow import PauliSumOp
 from qiskit.quantum_info.operators import Pauli
 
-from qiskit_nature.operators.second_quantization import FermionicOp
+from ..fermionic_op import FermionicOp
 
-from .fermionic_mapper import FermionicMapper
-from .qubit_mapper import QubitMapper
+from ...qubit_mapper import QubitMapper
 
 
-class ParityMapper(FermionicMapper):  # pylint: disable=missing-class-docstring
+class ParityMapper(QubitMapper):  # pylint: disable=missing-class-docstring
     def __init__(self):
         """The Parity fermion-to-qubit mapping.
